backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.0",
    "django>=5.2.1",
    "environs>=14.2.0",
    "newsapi-python>=0.2.7",
    "psycopg2-binary>=2.9.10",
    "pycountry>=24.6.1",
]

[dependency-groups]
dev = [
    "ruff>=0.11.11",
]
